+++ src/libponyc/ds/list.c
@@ -40,7 +40,7 @@ list_t* list_next(list_t* list)
   return list->next;
 }
 
-list_t* list_index(list_t* list, int index)
+list_t* list_index(list_t* list, size_t index)
 {
   if(index < 0)
     index = list_length(list) + index;
